+++ react-app/src/App.js
@@ -4,42 +4,21 @@ import ClipLoader from "react-spinners/ClipLoader";
 import { useCookies } from "react-cookie";
 
 import "./App.scss";
-import Landing from "./components/Landing/Landing.js";
-import Files from "./components/Files/Files.js";
+import Landing from "./pages/Landing/Landing.js";
+import Files from "./pages/Files/Files.js";
 import NavBar from "./components/NavBar/NavBar.js";
 import { UserContext } from "./contexts/UserContext";
-import { FilesContext } from "./contexts/FilesContext";
+import getProfile from "./utils/getProfile"
 
 function App() {
   const [isLoading, setIsLoading] = useState(true);
   const { userEmail, userID, updateUserContext } = useContext(UserContext);
-  const { files, updateFiles } = useContext(FilesContext);
   const [cookies, setCookie] = useCookies(["token"]);
 
-  console.log(userEmail);
-
   useEffect(() => {
-    const getProfile = async () => {
-      const res = await fetch("http://127.0.0.1:5000/profile", {
-        method: "GET",
-        headers: {
-          Jwt: cookies.token,
-        },
-      });
-
-      const apiResponse = await res.json();
-      const userProfile = apiResponse["user_profile"];
-      if (userProfile && userProfile.email && userProfile.id) {
-        updateUserContext({
-          userEmail: userProfile["email"],
-          userFirstName: userProfile["first_name"],
-          userLastName: userProfile["last_name"],
-          userID: userProfile["id"],
-        });
-      }
-    };
+   
     if (cookies.token) {
-      getProfile();
+      getProfile(cookies.token, updateUserContext);
     }
     setIsLoading(false);
   }, []);
