+++ react-app/src/components/CameraBar/CameraBar.js
@@ -1,0 +1,11 @@
+import "./CameraBar.scss";
+
+function CameraBar() {
+  return (
+    <div >
+      Camera bar here
+    </div>
+  );
+}
+
+export default CameraBar;
