+++ react-app/src/components/CameraBar/CameraBar.scss
