--- react-app/src/components/Files/Files.js
@@ -1,11 +1,0 @@
-import "./Files.scss";
-
-function Files() {
-  return (
-    <div className="landing-container">
-      Files here
-    </div>
-  );
-}
-
-export default Files;
