+++ react-app/src/pages/Files/Files.js
@@ -1,0 +1,10 @@
+import "./Files.scss";
+import CameraBar from '../../components/CameraBar/CameraBar'
+
+function Files() {
+  return <div>Files here
+    <CameraBar/>
+  </div>;
+}
+
+export default Files;
