+++ react-app/src/pages/Files/Files.scss
@@ -1,0 +1,2 @@
+@import '../../styles/shared.scss';
+
