+++ react-app/src/pages/Landing/Landing.scss
